requirements.txt
django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
javalang>=0.13.0
Pillow>=10.4
cryptography>=42.0
